requirement.txt
bs4
regex
pandas
request